% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediateDS.R
\name{mediateDS}
\alias{mediateDS}
\title{Causal Mediation Analysis}
\usage{
mediateDS(
  model.m,
  model.y,
  treat,
  mediator,
  boot,
  conf.level,
  robustSE,
  sims,
  seed,
  newobj
)
}
\arguments{
\item{model.m}{a string character, the name of a fitted model object for mediator.}

\item{model.y}{a string character, the name of a fitted model object for outcome.}

\item{treat}{a character string indicating the name of the treatment variable used
in the models. The treatment can be either binary (integer or a two-valued factor)
or continuous (numeric).}

\item{mediator}{a character string indicating the name of the mediator variable
used in the models.}

\item{boot}{a logical value. if 'FALSE' a quasi-Bayesian approximation is used for
confidence intervals; if 'TRUE' nonparametric bootstrap will be used. Default is 'FALSE'.}

\item{conf.level}{the level of the returned two-sided confidence intervals.}

\item{robustSE}{a logical value. If 'TRUE', heteroskedasticity-consistent standard
errors will be used in quasi-Bayesian simulations. Ignored if 'boot' is 'TRUE' or
neither 'model.m' nor 'model.y' has a method for vcovHC in the sandwich package.
Default is 'FALSE'.}

\item{sims}{a number of Monte Carlo draws for nonparametric bootstrap or
quasi-Bayesian approximation.}

\item{seed}{a number of a seed random number generator. Default value is NULL.}

\item{newobj}{a character string that provides the name for the output object
that is stored on the data servers. Default \code{med.out}.}
}
\value{
a summary table of the object of class 'mediate'.
}
\description{
This function is similar to R function \code{mediate} from the
\code{mediation} package.
}
\details{
The function 'mediate' is used to estimate various quantities for
causal mediation analysis, including average causal mediation effects
(indirect effect), average direct effects, proportions mediated, and total effect.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
