% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationTestDS.R
\name{mediationTestDS}
\alias{mediationTestDS}
\title{The Sobel mediation test}
\usage{
mediationTestDS(mv, iv, dv)
}
\arguments{
\item{mv}{a string character, the name of the mediator variable}

\item{iv}{a string character, the name of the independent variable}

\item{dv}{a string character, the name of the dependent variable}
}
\value{
A table showing the values of the test statistics (z-values) and the corresponding
p-values for the Sobel, Aroian and Goodman tests.
}
\description{
This function is similar to the R function \code{mediation.test} that is included
in the \code{bda} package. The function computes statistics and p-values for three versions of
Sobel mediation test: Sobel test, Aroian test and Goodman test.
}
\details{
To test whether a mediator carries the influence on an independent variable to a
dependent variable. Missing values will be automatically excluded with a warning.
}
\author{
Demetris Avraam for DataSHIELD Development Team
}
