% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neEffdecompDS.R
\name{neEffdecompDS}
\alias{neEffdecompDS}
\title{Linear hypotheses for natural effect models}
\usage{
neEffdecompDS(model.name)
}
\arguments{
\item{model.name}{the name of a fitted natural effect model object. This is the
object saved on the server-side by the \code{ds.neModel} function.}
}
\value{
a summary table of the object of class c("neLht", "glht") (see glht). The object
additionally inherits from class "neEffdecomp".
}
\description{
This function is similar to R function \code{neEffdecomp} from the
\code{medflex} package.
}
\details{
The function \code{ds.neEffdecomp} automatically extracts relevant causal
parameter estimates from a natural effect model.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
