% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neWeightDS.R
\name{neWeightDS}
\alias{neWeightDS}
\title{Expand the dataset and calculate ratio-of-mediator probability weights}
\usage{
neWeightDS(object)
}
\arguments{
\item{object}{a string character, the name of an object used to select a method.}
}
\value{
a data frame of class c("data.frame", "expData", "weightData") is assigned at
the server-side of each study.
}
\description{
This function is similar to R function \code{neWeight} from the
\code{mmedflex} package.
}
\details{
The function 'neWeight' both expands the data along hypothetical exposure
values and calculates ratio-of-mediator probability weights.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
